+++ src/Traits/Resources/HasMergeAdditional.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Php\Support\Laravel\Traits\Resources;
+
+trait HasMergeAdditional
+{
+    public function additional(array $data)
+    {
+        $this->additional = array_merge($this->additional, $data);
+
+        return $this;
+    }
+}