+++ tests/TestClasses/Models/TestModel.php
@@ -6,6 +6,7 @@ use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Foundation\Auth\User;
+use Php\Support\Laravel\Caster\GeoPoint;
 use Php\Support\Laravel\Caster\HasCasts;
 use Php\Support\Laravel\Tests\TestClasses\Entity\Params;
 use Php\Support\Laravel\Tests\TestClasses\Entity\Status;
@@ -16,6 +17,8 @@ use Php\Support\Laravel\Tests\TestClasses\Entity\Status;
  * @property boolean $enabled
  * @property string $title
  * @property Params $params
+ * @property Status $status
+ * @property GeoPoint $geo_point
  * @property array $config
  * @property string $str
  * @property string $str_empty
@@ -28,15 +31,18 @@ class TestModel extends Model
     use HasCasts;
 
     public $timestamps = false;
-    protected $keyType = 'uuid';
-    protected $table   = 'test_table';
+    protected $keyType = 'string';
+    protected $table = 'test_table';
 
     protected $fillable = [
         'params',
         'config',
         'str',
+        'status',
         'str_empty',
         'int',
+        'enabled',
+        'geo_point',
     ];
 
     protected $casts = [
@@ -45,8 +51,9 @@ class TestModel extends Model
         'status'    => Status::class,
         'config'    => 'array',
         'str'       => 'string',
-        'str_empty' => null,
+        'str_empty' => 'string',
         'int'       => 'int',
+        'geo_point' => GeoPoint::class,
     ];
 
     public function user(): BelongsTo