+++ tests/TestClasses/ServiceProviders/routes/back-api.php
@@ -1,0 +1,9 @@
+<?php
+
+use Illuminate\Support\Facades\Route;
+
+Route::prefix('admin')
+    ->group(
+        static function () {
+        }
+    );