+++ tests/database/factories/TestModelFactory.php
@@ -1,20 +1,29 @@
 <?php
 
-/** @var Factory $factory */
+namespace Php\Support\Laravel\Tests\Database\Factories;
 
-use Faker\Generator as Faker;
-use Illuminate\Database\Eloquent\Factory;
+use Illuminate\Database\Eloquent\Factories\Factory;
+use Php\Support\Laravel\Caster\GeoPoint;
 use Php\Support\Laravel\Tests\TestClasses\Entity\Status;
 use Php\Support\Laravel\Tests\TestClasses\Models\TestModel;
 
-$factory->define(
-    TestModel::class,
-    static function (Faker $faker) {
+class TestModelFactory extends Factory
+{
+    protected $model = TestModel::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition(): array
+    {
         return [
-            'title'   => $faker->sentence,
-            'str'     => $faker->title,
-            'enabled' => $faker->randomElement([true, false]),
-            'status'  => $faker->randomElement(Status::STATUSES),
+            'title'     => $this->faker->sentence,
+            'str'       => $this->faker->title,
+            'enabled'   => $this->faker->randomElement([true, false]),
+            'status'    => $this->faker->randomElement(Status::STATUSES),
+            'geo_point' => new GeoPoint($this->faker->randomFloat(), $this->faker->randomFloat()),
         ];
     }
-);
+}